classes/BLG361_Veritabanı Yönetim Sistemleri/2017_fall/mt_project/repo/requirements.txt
Flask
psycopg2
passlib
flask_login
Flask-WTF
wtforms_components
